// src/infraestructure/env/env-variables.enum.ts
export enum EnvVariables {

  /* DATABASE VARIABLES */
  DATABASE_TYPE = 'DATABASE_TYPE',
  DATABASE_PORT = 'DATABASE_PORT',
  DATABASE_HOST = 'DATABASE_HOST',
  DATABASE_USER = 'DATABASE_USER',
  DATABASE_PASSWORD = 'DATABASE_PASSWORD',
  DATABASE_NAME = 'DATABASE_NAME',

  TYPEORM_ENTITIES_DIR = 'TYPEORM_ENTITIES_DIR',
  TYPEORM_MIGRATIONS_DIR = 'TYPEORM_MIGRATIONS_DIR',
  TYPEORM_MIGRATIONS_TABLENAME = 'TYPEORM_MIGRATIONS_TABLENAME',
  TYPEORMCLI_MIGRATIONS_DIR = 'TYPEORMCLI_MIGRATIONS_DIR',

  /* APP VARIABLES */
  PORT = 'PORT',
  APPLICATION_CONTEXT_PATH = 'APPLICATION_CONTEXT_PATH',

  /* UPLASH VARIABLES */
  UPLASH_KEY = 'UPLASH_KEY',
  UPLASH_SECRET = 'UPLASH_SECRET',
  UPLASH_ID = 'UPLASH_ID',

  /* AWS VARIABLES */
  AWS_REGION = 'AWS_REGION',
  AWS_ACCESS_KEY_ID = 'AWS_ACCESS_KEY_ID',
  AWS_SECRET_ACCESS_KEY = 'AWS_SECRET_ACCESS_KEY',
  AWS_S3_BUCKET_NAME = 'AWS_S3_BUCKET_NAME',

  SECRET_KEYJWT = 'SECRET_KEYJWT',
  SECRET_RESET_PASSWORD = 'SECRET_RESET_PASSWORD',
}